+++ HISTORY.md
@@ -5,6 +5,8 @@ v1.5..10 (2022-09-09)
 - Updated
     - Improved the directory structure of the notes project.
     - Improve the way note headers are regenerated.
+- Added
+    - Create a pyinstaller spec file.
 
 v1.5..9 (2022-09-02)